Real-Time-Attendance-System/requirements-py38.txt
# Pinned, Python 3.8–compatible versions (recommended)
streamlit==1.31.0
opencv-python==4.7.0.72
face_recognition==1.3.0
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.3
Pillow==9.5.0
python-dotenv==1.0.1